Moved EventBus handling from the activity to the fragment
